+++ src/init.c
@@ -164,6 +164,8 @@ extern r_obj* ffi_vec_expand_grid(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_list_combine(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_compact_seq(r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_as_compact_condition(r_obj*);
+extern r_obj* ffi_vec_case_when(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*);
+extern r_obj* ffi_vec_replace_when(r_obj*, r_obj*, r_obj*, r_obj*);
 
 
 // Maturing
@@ -358,6 +360,8 @@ static const R_CallMethodDef CallEntries[] = {
   {"ffi_list_combine",                          (DL_FUNC) &ffi_list_combine, 11},
   {"ffi_compact_seq",                           (DL_FUNC) &ffi_compact_seq, 3},
   {"ffi_as_compact_condition",                  (DL_FUNC) &ffi_as_compact_condition, 1},
+  {"ffi_vec_case_when",                         (DL_FUNC) &ffi_vec_case_when, 7},
+  {"ffi_vec_replace_when",                      (DL_FUNC) &ffi_vec_replace_when, 4},
   {"ffi_exp_vec_cast",                          (DL_FUNC) &exp_vec_cast, 2},
   {NULL, NULL, 0}
 };